+++ engine/src/main/java/org/terasology/world/block/internal/BlockBuilder.java
@@ -146,9 +146,6 @@ public class BlockBuilder implements BlockBuilderHelper {
         block.setPenetrable(def.isPenetrable());
         block.setTargetable(def.isTargetable());
         block.setClimbable(def.isClimbable());
-        if (def.isInvisible()) {
-            block.setMeshGenerator(null);
-        }
         block.setTranslucent(def.isTranslucent());
         block.setDoubleSided(def.isDoubleSided());
         block.setShadowCasting(def.isShadowCasting());
