+++ modules/Core/src/main/java/org/terasology/logic/health/HealthAuthoritySystem.java
@@ -259,110 +259,6 @@ public class HealthAuthoritySystem extends BaseComponentSystem implements Update
         }
     }
 
-    // Debug commands
-
-    @Command(value = "kill", shortDescription = "Reduce the player's health to zero", runOnServer = true,
-            requiredPermission = PermissionManager.NO_PERMISSION)
-    public void killCommand(@Sender EntityRef client) {
-        ClientComponent clientComp = client.getComponent(ClientComponent.class);
-        HealthComponent health = clientComp.character.getComponent(HealthComponent.class);
-        if (health != null) {
-            clientComp.character.send(new DestroyEvent(clientComp.character, EntityRef.NULL, EngineDamageTypes.DIRECT.get()));
-        }
-    }
-
-    @Command(shortDescription = "Reduce the player's health by an amount", runOnServer = true)
-    public String damage(@Sender EntityRef client, @CommandParam("amount") int amount) {
-        ClientComponent clientComp = client.getComponent(ClientComponent.class);
-        clientComp.character.send(new DoDamageEvent(amount, EngineDamageTypes.DIRECT.get(), clientComp.character));
-
-        return "Inflicted damage of " + amount;
-    }
-
-    @Command(shortDescription = "Restores your health to max", runOnServer = true,
-            requiredPermission = PermissionManager.CHEAT_PERMISSION)
-    public String healthMax(@Sender EntityRef clientEntity) {
-        ClientComponent clientComp = clientEntity.getComponent(ClientComponent.class);
-        clientComp.character.send(new DoHealEvent(100000, clientComp.character));
-        return "Health restored";
-    }
-
-    @Command(shortDescription = "Restores your health by an amount", runOnServer = true,
-            requiredPermission = PermissionManager.CHEAT_PERMISSION)
-    public void health(@Sender EntityRef client, @CommandParam("amount") int amount) {
-        ClientComponent clientComp = client.getComponent(ClientComponent.class);
-        clientComp.character.send(new DoHealEvent(amount, clientComp.character));
-    }
-
-    @Command(shortDescription = "Set max health", runOnServer = true,
-            requiredPermission = PermissionManager.CHEAT_PERMISSION)
-    public String setMaxHealth(@Sender EntityRef client, @CommandParam("max") int max) {
-        ClientComponent clientComp = client.getComponent(ClientComponent.class);
-        HealthComponent health = clientComp.character.getComponent(HealthComponent.class);
-        if (health != null) {
-            doHeal(clientComp.character, health.maxHealth, clientComp.character);
-        }
-        return "Max health set to " + max;
-    }
-
-    @Command(shortDescription = "Set health regen rate", runOnServer = true,
-            requiredPermission = PermissionManager.CHEAT_PERMISSION)
-    public String setRegenRate(@Sender EntityRef client, @CommandParam("rate") float rate) {
-        ClientComponent clientComp = client.getComponent(ClientComponent.class);
-        HealthComponent health = clientComp.character.getComponent(HealthComponent.class);
-        if (health != null) {
-            health.regenRate = rate;
-            clientComp.character.saveComponent(health);
-        }
-        return "Set health regeneration rate to " + rate;
-    }
-
-    @Command(shortDescription = "Show your health")
-    public String showHealth(@Sender EntityRef client) {
-        ClientComponent clientComp = client.getComponent(ClientComponent.class);
-        HealthComponent health = clientComp.character.getComponent(HealthComponent.class);
-        if (health != null) {
-            return "Your health:" + health.currentHealth + " max:" + health.maxHealth + " regen:" + health.regenRate;
-        }
-        return "I guess you're dead?";
-    }
-
-
-    @Command(shortDescription = "Land without breaking a leg", runOnServer = true,
-            requiredPermission = PermissionManager.CHEAT_PERMISSION)
-    public String softLanding(@Sender EntityRef client) {
-        ClientComponent clientComp = client.getComponent(ClientComponent.class);
-        HealthComponent health = clientComp.character.getComponent(HealthComponent.class);
-        if (health != null) {
-            health.fallingDamageSpeedThreshold = 85f;
-            health.excessSpeedDamageMultiplier = 2f;
-            clientComp.character.saveComponent(health);
-
-            return "Soft landing mode activated";
-        }
-
-        return "";
-    }
-
-    @Command(shortDescription = "Restore default collision damage values", runOnServer = true,
-            requiredPermission = PermissionManager.CHEAT_PERMISSION)
-    public String restoreCollisionDamage(@Sender EntityRef client) {
-        ClientComponent clientComp = client.getComponent(ClientComponent.class);
-
-        Optional<Prefab> prefab = Assets.get(new ResourceUrn("engine:player"), Prefab.class);
-        HealthComponent healthDefault = prefab.get().getComponent(HealthComponent.class);
-        HealthComponent health = clientComp.character.getComponent(HealthComponent.class);
-        if (health != null && healthDefault != null) {
-            health.fallingDamageSpeedThreshold = healthDefault.fallingDamageSpeedThreshold;
-            health.horizontalDamageSpeedThreshold = healthDefault.horizontalDamageSpeedThreshold;
-            health.excessSpeedDamageMultiplier = healthDefault.excessSpeedDamageMultiplier;
-            clientComp.character.saveComponent(health);
-        }
-
-        return "Normal collision damage values restored";
-    }
-
-
     @ReceiveEvent
     public void onCrash(HorizontalCollisionEvent event, EntityRef entity, CharacterSoundComponent characterSounds, HealthComponent healthComponent) {
         Vector3f horizVelocity = new Vector3f(event.getVelocity());
